Use off_t for file size passed to vnode_create_vobject(). The former type, size_t, was causing truncation to 32 bits on i386, which immediately led to undersizing of VM objects backed by files >4GB.  In particular, sendfile(2) was broken for such files.
PR:		kern/92243
MFC after:	5 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@155177 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/vm/vnode_pager.c
@@ -97,7 +97,7 @@ int vnode_pbuf_freecnt;
 
 /* Create the VM system backing object for this vnode */
 int
-vnode_create_vobject(struct vnode *vp, size_t isize, struct thread *td)
+vnode_create_vobject(struct vnode *vp, off_t isize, struct thread *td)
 {
 	vm_object_t object;
 	vm_ooffset_t size = isize;
